requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
faster-whisper==1.1.1
--extra-index-url https://download.pytorch.org/whl/cu124
ctranslate2==4.6.0
torch==2.8.0
torchaudio==2.8.0
python-multipart==0.0.7
python-dotenv==1.0.0
sounddevice==0.4.6
numpy>=1.26.0
requests==2.32.3
pynput==1.7.8
scipy==1.13.1
pyautogui==0.9.54
Pillow==11.1.0
chime==0.7.0
desktop-notifier==4.0.0

# macOS menu bar support (optional)
rumps==0.4.0; sys.platform == "darwin"

# Build dependencies (optional, for creating standalone app)
pyinstaller==6.11.0; sys.platform == "darwin"
